Removed unnecessary unrelated code in example file
A teacher just tested the code in front of his students. 

We thought it was just copy and paste from the pub.dev website but as it turns out there's an unrelated import called 'routes.dart' which wasted quite a bit of time to remove. This was the final modification, hopefully to avoid future embarassments.